Add nodemon configuration
